% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.dataset.R
\name{dataset}
\alias{dataset}
\title{Create itol dataset file}
\usage{
dataset(TYPE, DATA, FILE, ...)
}
\arguments{
\item{TYPE}{dataset type}

\item{DATA}{dataframe with the data}

\item{FILE}{output file name}

\item{...}{fields to add to the metadata block}
}
\description{
Create itol dataset file
}
