% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hmmer.domtblout.R
\name{read.hmmer.domtblout}
\alias{read.hmmer.domtblout}
\title{Read hmmer domtblout file}
\usage{
read.hmmer.domtblout(fname)
}
\arguments{
\item{fname}{file name}
}
\description{
Read hmmer domtblout file
}
