% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.itol.dataset.R
\name{read.itol.dataset}
\alias{read.itol.dataset}
\title{Read itol dataset file}
\usage{
read.itol.dataset(FILE)
}
\arguments{
\item{FILE}{output file name}

\item{TYPE}{dataset type}

\item{DATA}{dataframe with the data}

\item{...}{fields to add to the metadata block}
}
\description{
Read itol dataset file
}
