% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.itol.dataset.R
\name{write.itol.dataset}
\alias{write.itol.dataset}
\title{Create itol dataset file}
\usage{
write.itol.dataset(TYPE, DATA, FILE, ...)
}
\arguments{
\item{TYPE}{dataset type}

\item{DATA}{dataframe with the data}

\item{FILE}{output file name}

\item{...}{fields to add to the metadata block}
}
\description{
Create itol dataset file
}
